// web/generated/schemas/validationErrorResponseDto.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * API
 * Upload and visualize data
 * OpenAPI spec version: 1.0
 */
import type { ValidationErrorItemDto } from './validationErrorItemDto';

export interface ValidationErrorResponseDto {
  errors: ValidationErrorItemDto[];
  message: string;
  statusCode: number;
}